Resolved g++ warning about narrowing conversion

+++ idg-lib/src/CPU/Reference/kernels/KernelFFT.cpp
@@ -19,7 +19,7 @@ void kernel_fft(
     int rank = 2;
 
     // For grids of size*size elements
-    int n[] = {size, size};
+    long n[] = {size, size};
 
     // Set stride
     int istride = 1;
